Updated test cases for AssetPrice and FetchAsJson
